절반 정도 코드 분리
