+++ data/app_variables.json
@@ -1,6 +1,6 @@
 {
     "_README" : "DO NOT MODIFY THIS FILE.",
-    "PATH_OAUTHSETTINGS" : "data/oauthsettings.json",
-    "PATH_SET_JSON" : "data/set.json",
-    "PATH_DATA_IMAGES" : "data/images"
+    "path_oauth_setting" : "data/oauthsettings.json",
+    "path_set_json" : "data/set.json",
+    "path_data_image" : "data/images"
 }
